% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.r
\name{calculate_active_users}
\alias{calculate_active_users}
\title{Calculate the number of active users in a set of date ranges.}
\usage{
calculate_active_users(sesh_dur_data, range_beginning)
}
\arguments{
\item{sesh_dur_data}{A dataframe of user_ids and the dates when they took a
platform action.}

\item{range_beginning}{A data.table with a single column of dates which 
denote the first days of each date range.}
}
\value{
A data frame showing how many active users there were in each date
range.
}
\description{
Calculate the number of active users in a set of date ranges.
}

