% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.r
\name{calculate_DAU_MAU_ratio}
\alias{calculate_DAU_MAU_ratio}
\title{Calculate DAU to MAU ratio by month.}
\usage{
calculate_DAU_MAU_ratio(MAU_count_by_month, DAU_count_by_day)
}
\arguments{
\item{MAU_count_by_month}{A data.table: (month_beginning MAU_count)}

\item{DAU_count_by_day}{A data.table: (day DAU_count)}
}
\value{
A data frame showing the average DAU to MAU ratio for each month
specified by months_to_calc.
}
\description{
Calculate DAU to MAU ratio by month.
}

