% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.r
\name{calculate_overall_ratio}
\alias{calculate_overall_ratio}
\title{Calculate overall DAU to MAU ratio for a time interval.}
\usage{
calculate_overall_ratio(months_beginning, max_date = Sys.Date(),
  sesh_dur_data = sesh_dur_date)
}
\arguments{
\item{months_beginning}{A data.table (range_beginning_date) of months to
include in the calculation.}

\item{max_date}{The largest date to use in the calculation. Any of the
range_beginning_dates in months_beginning which is less than one month
before max_date is dropped.}

\item{sesh_dur_data}{A data.table (user_id, platform_action_date).}
}
\value{
A number representing the overall DAU/MAU ratio for the time
interval (min_date, max_date). The ratio is only calculated for complete
months within the time interval. So for example, (jan 3, march 20) gets
rounded to (feb 1, feb 28).
}
\description{
Calculate overall DAU to MAU ratio for a time interval.
}

