% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{classify_1d7s}
\alias{classify_1d7s}
\title{Classify users as 1d7 or regular.}
\usage{
classify_1d7s(user_pa_datetime_relative)
}
\arguments{
\item{user_pa_datetime_relative}{A data frame: 
(user_id, relative_datetime, relative_date). The
result of calling find_relative_pa_datetimes.}
}
\value{
A data frame: (user_id, 1d7). The field 1d7 is logical; true if the
user returned to Gloo in their first week after signup and false otherwise.
}
\description{
Classify users as 1d7 or regular.
}

