% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{count_days_of_activity_week1}
\alias{count_days_of_activity_week1}
\title{Count the number of distinct days of activity each user had in their first
week.}
\usage{
count_days_of_activity_week1(user_pa_datetime_relative)
}
\arguments{
\item{user_pa_datetime_relative}{A data frame: 
(user_id, relative_datetime, relative_date). The
result of calling find_relative_pa_datetimes.}
}
\value{
A data frame: (user_id, user_class). The field 'class' is numeric; it
corresponds to the number of days of activity the user had in their first
week.
}
\description{
Count the number of distinct days of activity each user had in their first
week.
}

