% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.r
\name{count_signups}
\alias{count_signups}
\title{Count signups per week (as total, and as perentage of signups).}
\usage{
count_signups(relative_pa_datetimes, user_oneD7,
  min_date = as.Date("2016-01-01"),
  max_date = max(date_ranges[["days"]]$range_beginning_date))
}
\arguments{
\item{relative_pa_datetimes}{A data frame: (user_id, datetime) giving the
moments when a user took a platform action.}

\item{user_oneD7}{A data frame: (user_id oneD7) that tells us which users
are 1D7s.}
}
\value{
A data frame: (week_beginning, number_of_signups)
}
\description{
Count signups per week (as total, and as perentage of signups).
}

