% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.r
\name{define_date_ranges}
\alias{define_date_ranges}
\title{Return a list of date intervals relative to the current date.}
\usage{
define_date_ranges(base_date = Sys.Date())
}
\arguments{
\item{base_date}{A date object.}
}
\value{
interval_list A list of (date_interval, functions_to_apply)  pairs.
 See standardmetrics::interval_list for an example of how to format this list.
}
\description{
Return a list of date intervals relative to the current date.
}

