% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.r
\name{extract_user_pa_date}
\alias{extract_user_pa_date}
\title{Download user_platform_action_date data from Looker, and format date column
correctly.}
\usage{
extract_user_pa_date()
}
\value{
A data frame with a user_id column and a
platform_action_date_column.
}
\description{
Download user_platform_action_date data from Looker, and format date column
correctly.
}

