% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.r
\name{extract_user_pa_datetime}
\alias{extract_user_pa_datetime}
\title{Download session_datetime data from Looker, and format date and time columns
correctly.}
\usage{
extract_user_pa_datetime()
}
\value{
A data frame with columns user_id, date, time.
}
\description{
Download session_datetime data from Looker, and format date and time columns
correctly.
}

