% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_data.r
\name{find_relative_pa_datetimes}
\alias{find_relative_pa_datetimes}
\title{Convert platform action datetimes to difftimes relative to first platform
action, and relative to first day.}
\usage{
find_relative_pa_datetimes(user_platform_action_datetime)
}
\arguments{
\item{user_platform_action_datetime}{A data frame: (user_id, datetime) 
where datetime is a chron
object representing the moment a platform action was taken.}
}
\value{
A data frame: (user_id, relative_time, relative_date) where
relative_* are difftime objects. relative_time represents the amount of 
seconds that have
passed since their first platform action. relative_date represents the
amount of days that have passed since the first day of their platform
action.
}
\description{
Convert platform action datetimes to difftimes relative to first platform
action, and relative to first day.
}

