% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.r
\name{get_user_retention_data}
\alias{get_user_retention_data}
\title{Get user retention data (1d7s vs all).}
\usage{
get_user_retention_data(relative_pa_datetimes, user_oneD7,
  max_date = max(sesh_dur_date$platform_action_date))
}
\arguments{
\item{relative_pa_datetimes}{A data frame: (user_id, datetime) giving the
moments when a user took a platform action.}

\item{user_oneD7}{A data frame: (user_id oneD7) that tells us which users
are 1D7s.}

\item{max_date}{The most recent date of activity that appears in
sesh_dur_date.}
}
\value{
A data frame: (weeks_since_signup, pct_users_active)
}
\description{
Get user retention data (1d7s vs all).
}

