% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.r
\name{loop_calculate_active_users}
\alias{loop_calculate_active_users}
\title{Loop calculate_active_users over a list of date ranges.}
\usage{
loop_calculate_active_users(date_ranges, ...)
}
\arguments{
\item{date_ranges}{A list of date ranges, of the form generated by
standardmetrics::define_date_ranges.}
}
\value{
A list of data frames that show the number of active users that fall
within each date range.
}
\description{
Loop calculate_active_users over a list of date ranges.
}

